combine canvas rules into a single annotations rule
