+++ rules/dv-rule-annotations-001/tests/IncorrectAnnotationColor.expect.json
@@ -1,0 +1,51 @@
+{
+  "name": "PingOne DaVinci Linter",
+  "pass": false,
+  "errorCount": 2,
+  "rulePackResults": [
+    {
+      "name": "linter-davinci-base-pack",
+      "pass": false,
+      "errorCount": 2,
+      "lintResults": [
+        {
+          "flowId": "d493071599c7996c28591f29b3578f2e",
+          "flowName": "Test Annotation Colors",
+          "pass": false,
+          "errorCount": 2,
+          "errors": [
+            "Annotation color is not in palette [#3030aeff] (4pmrtt0pi7)",
+            "Each flow should contain a title annotation node with a background color of #4462ed"
+          ],
+          "rulesApplied": [
+            "dv-rule-annotations-001"
+          ],
+          "ruleResults": [
+            {
+              "ruleId": "dv-rule-annotations-001",
+              "pass": false,
+              "errorCount": 2,
+              "errors": [
+                {
+                  "code": "dv-bp-annotation-001",
+                  "message": "Annotation color is not in palette [#3030aeff] (4pmrtt0pi7)",
+                  "type": "best-practice",
+                  "recommendation": "Use recommended CSS settings for Ping developed flows",
+                  "nodeId": "4pmrtt0pi7"
+                },
+                {
+                  "code": "dv-bp-missing-title-annotation-001",
+                  "message": "Each flow should contain a title annotation node with a background color of #4462ed",
+                  "type": "best-practice",
+                  "recommendation": "Add a title annotation node"
+                }
+              ]
+            }
+          ],
+          "rulesIgnored": []
+        }
+      ],
+      "rulesIgnored": true
+    }
+  ]
+}
