+++ rules/dv-rule-annotations-001/tests/IncorrectAnnotationFont.expect.json
@@ -1,30 +1,38 @@
 {
   "name": "PingOne DaVinci Linter",
   "pass": false,
-  "errorCount": 1,
+  "errorCount": 2,
   "rulePackResults": [
     {
+      "name": "linter-davinci-base-pack",
       "pass": false,
-      "errorCount": 1,
+      "errorCount": 2,
       "lintResults": [
         {
           "flowId": "d493071599c7996c28591f29b3578f2e",
           "flowName": "Test Annotation Colors",
           "pass": false,
-          "errorCount": 1,
+          "errorCount": 2,
           "errors": [
+            "Annotation color is not in palette [#e7e7e7ff] (4pmrtt0pi7)",
             "Annotation font is not in sans-serif [verdana] (4pmrtt0pi7)"
           ],
           "rulesApplied": [
-            "dv-rule-canvas-002"
+            "dv-rule-annotations-001"
           ],
           "ruleResults": [
             {
-              "ruleId": "dv-rule-canvas-002",
-              "ruleDescription": "Verify and clean font family for node annotations",
+              "ruleId": "dv-rule-annotations-001",
               "pass": false,
-              "errorCount": 1,
+              "errorCount": 2,
               "errors": [
+                {
+                  "code": "dv-bp-annotation-001",
+                  "message": "Annotation color is not in palette [#e7e7e7ff] (4pmrtt0pi7)",
+                  "type": "best-practice",
+                  "recommendation": "Use recommended CSS settings for Ping developed flows",
+                  "nodeId": "4pmrtt0pi7"
+                },
                 {
                   "code": "dv-bp-annotation-002",
                   "message": "Annotation font is not in sans-serif [verdana] (4pmrtt0pi7)",
