+++ rules/dv-rule-annotations-001/tests/MissingAnnotations.expect.json
@@ -1,30 +1,37 @@
 {
   "name": "PingOne DaVinci Linter",
   "pass": false,
-  "errorCount": 1,
+  "errorCount": 2,
   "rulePackResults": [
     {
+      "name": "linter-davinci-base-pack",
       "pass": false,
-      "errorCount": 1,
+      "errorCount": 2,
       "lintResults": [
         {
           "flowId": "01fe43ef0bf69d071d13dc9ee1a6d298",
           "flowName": "Missing Annotations",
           "pass": false,
-          "errorCount": 1,
+          "errorCount": 2,
           "errors": [
+            "Each flow should contain a title annotation node with a background color of #4462ed",
             "Flow contains no annotations"
           ],
           "rulesApplied": [
-            "dv-rule-canvas-005"
+            "dv-rule-annotations-001"
           ],
           "ruleResults": [
             {
-              "ruleId": "dv-rule-canvas-005",
-              "ruleDescription": "Verify flow contains annotations",
+              "ruleId": "dv-rule-annotations-001",
               "pass": false,
-              "errorCount": 1,
+              "errorCount": 2,
               "errors": [
+                {
+                  "code": "dv-bp-missing-title-annotation-001",
+                  "message": "Each flow should contain a title annotation node with a background color of #4462ed",
+                  "type": "best-practice",
+                  "recommendation": "Add a title annotation node"
+                },
                 {
                   "code": "dv-bp-annotation-003",
                   "message": "Flow contains no annotations",
