+++ rules/dv-rule-annotations-001/tests/MissingTitleAnnotation.expect.json
@@ -1,35 +1,43 @@
 {
   "name": "PingOne DaVinci Linter",
   "pass": false,
-  "errorCount": 1,
+  "errorCount": 2,
   "rulePackResults": [
     {
+      "name": "linter-davinci-base-pack",
       "pass": false,
-      "errorCount": 1,
+      "errorCount": 2,
       "lintResults": [
         {
           "flowId": "bd15b4d79336dcd52897e070738ad233",
           "flowName": "Missing Title Annotation",
           "pass": false,
-          "errorCount": 1,
+          "errorCount": 2,
           "errors": [
+            "Annotation color is not in palette [#e7e7e7ff] (2elun9ixpr)",
             "Each flow should contain a title annotation node with a background color of #4462ed"
           ],
           "rulesApplied": [
-            "dv-rule-canvas-003"
+            "dv-rule-annotations-001"
           ],
           "ruleResults": [
             {
-              "ruleId": "dv-rule-canvas-003",
-              "ruleDescription": "Verify flow contains title annotation",
+              "ruleId": "dv-rule-annotations-001",
               "pass": false,
-              "errorCount": 1,
+              "errorCount": 2,
               "errors": [
+                {
+                  "code": "dv-bp-annotation-001",
+                  "message": "Annotation color is not in palette [#e7e7e7ff] (2elun9ixpr)",
+                  "type": "best-practice",
+                  "recommendation": "Use recommended CSS settings for Ping developed flows",
+                  "nodeId": "2elun9ixpr"
+                },
                 {
                   "code": "dv-bp-missing-title-annotation-001",
                   "message": "Each flow should contain a title annotation node with a background color of #4462ed",
                   "type": "best-practice",
-                  "recommendation": "Add a title annotation node (%)"
+                  "recommendation": "Add a title annotation node"
                 }
               ]
             }
