+++ rules/dv-rule-annotations-001/tests/TitleNotOnTop.expect.json
@@ -4,33 +4,33 @@
   "errorCount": 1,
   "rulePackResults": [
     {
+      "name": "linter-davinci-base-pack",
       "pass": false,
       "errorCount": 1,
       "lintResults": [
         {
-          "flowId": "d493071599c7996c28591f29b3578f2e",
-          "flowName": "Test Annotation Colors",
+          "flowId": "c6394f3b116bf75044b8af2b6b5c2104",
+          "flowName": "Title Not On Top",
           "pass": false,
           "errorCount": 1,
           "errors": [
-            "Annotation color is not in palette [#3030aeff] (4pmrtt0pi7)"
+            "The topmost node in your flow should be a title annotation (v5t8nizy0z)"
           ],
           "rulesApplied": [
-            "dv-rule-canvas-001"
+            "dv-rule-annotations-001"
           ],
           "ruleResults": [
             {
-              "ruleId": "dv-rule-canvas-001",
-              "ruleDescription": "Verify colors for node annotations",
+              "ruleId": "dv-rule-annotations-001",
               "pass": false,
               "errorCount": 1,
               "errors": [
                 {
-                  "code": "dv-bp-annotation-001",
-                  "message": "Annotation color is not in palette [#3030aeff] (4pmrtt0pi7)",
+                  "code": "dv-bp-title-not-on-top-001",
+                  "message": "The topmost node in your flow should be a title annotation (v5t8nizy0z)",
                   "type": "best-practice",
-                  "recommendation": "Use recommended CSS settings for Ping developed flows",
-                  "nodeId": "4pmrtt0pi7"
+                  "recommendation": "Position title annotation at top of the flow",
+                  "nodeId": "v5t8nizy0z"
                 }
               ]
             }
