--- rules/dv-rule-canvas-004/tests/TitleNotOnTop.expect.json
@@ -1,44 +1,0 @@
-{
-  "name": "PingOne DaVinci Linter",
-  "pass": false,
-  "errorCount": 1,
-  "rulePackResults": [
-    {
-      "pass": false,
-      "errorCount": 1,
-      "lintResults": [
-        {
-          "flowId": "c6394f3b116bf75044b8af2b6b5c2104",
-          "flowName": "Title Not On Top",
-          "pass": false,
-          "errorCount": 1,
-          "errors": [
-            "The topmost node in your flow should be a title annotation (v5t8nizy0z)"
-          ],
-          "rulesApplied": [
-            "dv-rule-canvas-004"
-          ],
-          "ruleResults": [
-            {
-              "ruleId": "dv-rule-canvas-004",
-              "ruleDescription": "Verify Title Annotation is on top",
-              "pass": false,
-              "errorCount": 1,
-              "errors": [
-                {
-                  "code": "dv-bp-title-not-on-top-001",
-                  "message": "The topmost node in your flow should be a title annotation (v5t8nizy0z)",
-                  "type": "best-practice",
-                  "recommendation": "Position title annotation at top of the flow",
-                  "nodeId": "v5t8nizy0z"
-                }
-              ]
-            }
-          ],
-          "rulesIgnored": []
-        }
-      ],
-      "rulesIgnored": true
-    }
-  ]
-}
